Implement minimum routing
