+++ src/About.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from 'react';
+import { Link } from 'react-router-dom';
+
+const About: FC = () => {
+  return (
+    <React.Fragment>
+      <h1>Aboutページです</h1>
+    </React.Fragment>
+  )
+}
+
+export default About;
